Do esbuild and tailwind in nix's mix release
This was painful. To make things reproducible without using the mix
package's download mechanisms, we have to create derivations for the
esbuild and tailwind binaries. Those get used in the preConfigure stage
of mixRelease.

But it's not as simple as calling mix assets.deploy. At that stage,
there are no mix dependencies, so I've copy+pasted what that task does
for now. Had to also copy in the js dependencies from the mix deps in
order for the esbuild to succeed.

For now, the hosting project still uses the hex dependencies for esbuild
and tailwind, with the binaries provided by nix. Could think about
transitioning fully away from the hex dependencies, or to fix how this
stuff happens upstream somehow.